+++ Socolin.RabbitMQ.Client/Options/Consumer/ConsumerOptionsBuilder.cs
@@ -4,6 +4,7 @@ using System.Collections.ObjectModel;
 using Socolin.RabbitMQ.Client.Exceptions;
 using Socolin.RabbitMQ.Client.Pipes.Consumer;
 using Socolin.RabbitMQ.Client.Pipes.Consumer.Builders;
+using Socolin.RabbitMQ.Client.Pipes.Consumer.Context;
 
 namespace Socolin.RabbitMQ.Client.Options.Consumer
 {
@@ -20,6 +21,13 @@ namespace Socolin.RabbitMQ.Client.Options.Consumer
 			return this;
 		}
 
+		/// <summary>
+		/// Configure the deserializer to use for a given ContentType. If a message's ContentType does not match
+		/// any deserializer, it use the default one
+		/// </summary>
+		/// <param name="deserializer"></param>
+		/// <param name="contentType"></param>
+		/// <returns></returns>
 		public ConsumerOptionsBuilder<T> WithDeSerializer(Func<ReadOnlyMemory<byte>, T> deserializer, string contentType)
 		{
 			_deserializers[contentType] = deserializer;
@@ -32,6 +40,10 @@ namespace Socolin.RabbitMQ.Client.Options.Consumer
 			return this;
 		}
 
+		/// <summary>
+		/// Configure the pipeline to Ack successfully processed messages (no exception has been thrown when executing
+		/// <see cref="IConsumerPipeContext{T}.MessageProcessor">MessageProcessor</see>) and to reject message on error.
+		/// </summary>
 		public ConsumerOptionsBuilder<T> WithSimpleMessageAck()
 		{
 			_messageAcknowledgmentPipeBuilder = new DefaultConsumerPipeBuilder<T>(() =>
@@ -40,6 +52,14 @@ namespace Socolin.RabbitMQ.Client.Options.Consumer
 			return this;
 		}
 
+		/// <summary>
+		/// Configure the pipeline to Ack successfully processed messages (no exception has been thrown when executing
+		/// <see cref="IConsumerPipeContext{T}.MessageProcessor">MessageProcessor</see>) and to retry message when
+		/// processing fail. On failure the message is immediately requeue.
+		/// If the message has been retried too many time, it will be rejected.
+		/// </summary>
+		/// <param name="maxRetryCount"></param>
+		/// <param name="retryCountHeaderName">The name of the header used to keep track of the number of retry. Default="RetryCount"</param>
 		public ConsumerOptionsBuilder<T> WithFastRetryMessageAck(int maxRetryCount, string? retryCountHeaderName = null)
 		{
 			_messageAcknowledgmentPipeBuilder = new DefaultConsumerPipeBuilder<T>(() =>
@@ -48,6 +68,32 @@ namespace Socolin.RabbitMQ.Client.Options.Consumer
 			return this;
 		}
 
+		/// <summary>
+		/// Configure the pipeline to Ack successfully processed messages (no exception has been thrown when executing
+		/// <see cref="IConsumerPipeContext{T}.MessageProcessor">MessageProcessor</see>) and to retry message when
+		/// processing fail. On failure the message is requeue with an expiration time in another queue that *must*
+		/// be created with the original queue as dead letter queue. This can be done by calling
+		/// <see cref="DelayedRetryMessageAcknowledgementPipe{T}.CreateDelayQueueAsync"/>
+		/// If the message has been retried too many time, it will be rejected.
+		/// </summary>
+		/// <param name="retryDelays"></param>
+		/// <param name="delayedMessagesQueueName">The name of the queue used to keep message until they expire and
+		/// they are re-queued to the original queue. Default : queueName + "-Delayed"</param>
+		/// <param name="retryCountHeaderName">The name of the header used to keep track of the number of retry.
+		/// Default : "DelayedRetryCount"</param>
+		public ConsumerOptionsBuilder<T> WithDelayedRetryMessageAck(TimeSpan[] retryDelays, string? delayedMessagesQueueName = null, string? retryCountHeaderName = null)
+		{
+			_messageAcknowledgmentPipeBuilder = new DefaultConsumerPipeBuilder<T>(() =>
+				new DelayedRetryMessageAcknowledgementPipe<T>(retryDelays, delayedMessagesQueueName, retryCountHeaderName)
+			);
+			return this;
+		}
+
+		/// <summary>
+		/// Configure the pipe responsible to Ack/Reject the messages
+		/// </summary>
+		/// <param name="pipeBuilder"></param>
+		/// <returns></returns>
 		public ConsumerOptionsBuilder<T> WitheMessageAck(IConsumerPipeBuilder<T> pipeBuilder)
 		{
 			_messageAcknowledgmentPipeBuilder = pipeBuilder;